rename birthdays to home
